Improve sensitivity of last two integration tests
Those tests were only checking that the output was nonempty, so they
would pretty much never detect anything fishy. We could go further but
it's nontrivial because the output is nondeterministic (because of
temporary files) and it's unclear if we want to build complex regexes in
the test suite right now. In the meantime, here's a basic improvement.