human_readable flag (-h) functionality enabled
This re-enables a quick, easy to type command line flag that `ls` users
are familiar with.

In the event a user has changed the default configuration of `lsd` to
use a size setting of something other than `default`, this lets a user
use the historically established CLI flag (`-h`) to enable the human
readable option without having to type `--size=default` on the command
line.