update various dependencies (#938)
I'm the maintainer of lsd for Fedora Linux, and some outdated
dependencies are making maintenance increasingly difficult.

- deps: update chrono-humanize to v0.2

We have this patched for almost two years in Fedora and it has not
caused issues.

- deps: update assert_cmd to v2

Same here, this patch has been in the Fedora package for a while.

- deps: update sys-locale to v0.3

This dependency seems to have been added recently, not sure why an old
version was chosen.

- deps: update vsort to v0.2

Same here, this was added recently but 0.1 was used instead of 0.2, not
sure why.

- deps: update git2 to v0.18

Using old versions of git2 is not a good idea, since the bundled libgit2
C library often has CVE issues.

- deps: migrate from users to uzers

The "users" crate is unmaintained. The "uzers" crate is an
API-compatible fork that also fixes some bugs and security issues.

- deps: update serial_test to v2

The current dependency (v0.5) is **reeeeeally** old. Not sure why this
was never updated.

- deps: update predicates to v3

Same here, predicates v1 is **reaally** old.

- deps: allow newer versions of url, wild, and xdg crates

Not sure why strange `x.0.*` style dependencies were used here. It's
holding back various updates for both url and xdg crates, and makes
maintaining lsd in Fedora more difficult. We have built lsd against the
latest versions of all three crates forever, and it has not caused
issues.