update various dependencies
- update dirs from v4 to v5 (no code changes)
- update term_grid from v0.1 to v0.2 (small code changes)
- update terminal_size from v0.1 to v0.3 (no code changes)
- update lscolors from v0.15 to v0.16 (no code changes)
- update serde_yaml from v0.8 to v0.9 (no code changes)

The term_grid update required small changes to the use of
`term_grid::Cell` because with v0.2, it is possible to specify the
alignment within the cell. Adding `Alignment::Left` should preserve
the current behaviour.

This commit also includes the changes generated by "cargo update",
which seems not to have been run in a long time, causing very old
versions to be pulled in.