Fix tests with `NO_COLOR=1` set
Fix tests to force internal coloring independently of `NO_COLOR`
in the calling environment, to ensure that they pass.  This is needed
when `NO_COLOR` is set globally in the environment, e.g. when collecting
plain text test logs.

Technically, `crossterm::style::force_color_output()` leaks to other
tests.  However, there should be no harm in setting it globally,
and the library does not seem to provide a trivial way to scope
the change.  Just call it in both places where it is actually necessary,
to ensure that things work correctly also when only parts of the test
suite are run.

Fixes #1136

Signed-off-by: Michał Górny <mgorny@gentoo.org>