docs(README): update documentation and add sample files
This commit contains a huge update to the project documentation in hopes of making the process of installing, using, and customizing lsd friendlier for first time users and novices. The overall changes in this commit entail improving the readability of the documentation, ensuring the documentation is easier to grok, fixing broken  markdown references, and adding more context to the process of customizing lsd.

* Add markdown alerts
* Fix incorrect header depths
* Update default branch to reflect current status of the project
* Update installation section
* Clarify installation prerequisites
* Move alias section to installation instructions as an optional step
* Rename configuration section to customization section
* Remove customization related file contents
* Add sample files to remove friction for first time users
* Split out color section into its own dedicated document
* Fix broken color previews and add descriptive name to alt text
* Clean up FAQ section
* Update Github pull request template